remove sdl2, move main to binary and unify the engines
