+++ bindings/main.py
@@ -2,11 +2,11 @@ from gg import Game, Sprite
 
 
 def setup(g: Game) -> None:
-
-    s = Sprite(id="tank", rect=(10, 10, 10, 10))
-    for i in range(100):
-        tank_py = g.create_entity()
-        tank_py.add(s)
+    print("Hello from python")
+    # s = Sprite(id="tank", rect=(10, 10, 10, 10))
+    # for i in range(100):
+    #     tank_py = g.create_entity()
+    #     tank_py.add(s)
 
 
 def update() -> None:
