+++ src/Graphics/index_buffer.h
@@ -8,6 +8,6 @@ class IndexBuffer {
   public:
     IndexBuffer(const void *data, unsigned int size);
     ~IndexBuffer();
-    void Bind();
+    void Bind() const;
     void Unbind();
 };
