+++ src/Graphics/sprite_renderer.cpp
@@ -1,14 +1,16 @@
 #include "sprite_renderer.h"
 
-#include "Graphics/index_buffer.h"
 #include "Graphics/texture.h"
 #include "Utils/logger.h"
+#include <glad/glad.h>
 
 #include <glm/ext/matrix_transform.hpp>
 
-SpriteRenderer::SpriteRenderer() : m_Shader(Shader({"./assets/shaders/color.frag", "./assets/shaders/default.vert"})) {
+SpriteRenderer::SpriteRenderer()
+    : m_Ebo(nullptr, 0), m_Shader(Shader({"./assets/shaders/color.frag", "./assets/shaders/default.vert"})) {
+  Logger::info("Sprite renderer spawn");
   VertexBuffer vbo(quad, sizeof(quad));
-  IndexBuffer ebo(indices, sizeof(indices));
+  this->m_Ebo = IndexBuffer(indices, sizeof(indices));
   VertexBufferLayout layout;
 
   layout.AddAttribute<float>("position", 2);
@@ -21,6 +23,9 @@ SpriteRenderer::SpriteRenderer() : m_Shader(Shader({"./assets/shaders/color.frag
 };
 
 void SpriteRenderer::render(const Texture *texture, const glm::vec2 &position) const {
+  m_Vao.Bind();
+  m_Ebo.Bind();
+
   if (texture == nullptr) {
     Logger::err("Texture not found");
     return;
@@ -29,5 +34,9 @@ void SpriteRenderer::render(const Texture *texture, const glm::vec2 &position) c
   glm::mat4 model(1.0f);
   model = glm::translate(model, glm::vec3(position, 0));
   model = glm::scale(model, glm::vec3(texture->width, texture->height, 0));
+
+  m_Shader.Bind();
   m_Shader.SetMat4("model", model);
+
+  glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, 0);
 }
